add few tests for the iterator as well
